Seperate algorithms into seperate files
