+++ PrimeCSharp/solution_4/Program.cs
@@ -13,34 +13,16 @@ namespace FastSieve
 			Console.WriteLine("hamarb123's C# implementation: Started");
 
 			//this could be an array that could be precalculated and then stored in code, but I feel that this is more readable
-			Sieve.InitializePopcountByte();
+			OptimizedUnmanaged1BitSieve.InitializePopcountByte();
 
 			//this could be an array that could be precalculated and then stored in code, but I feel that this is more readable
-			Sieve.InitializeNextPrime();
+			OptimizedUnmanaged1BitSieve.InitializeNextPrime();
 
 			//Jit each method by calling so metrics aren't scewed by first-time run of method:
 			{
+				for (int i = 1; i <= 7; i++)
 				{
-					using Sieve sieve = new Sieve();
-					sieve.Initialize(1, 0);
-					sieve.RunSieve();
-					sieve.CountPrimes();
-				}
-				{
-					using Sieve sieve = new Sieve();
-					sieve.Initialize(2, 0);
-					sieve.RunSieve();
-					sieve.CountPrimes();
-				}
-				{
-					using Sieve sieve = new Sieve();
-					sieve.Initialize(4, 0);
-					sieve.RunSieve();
-					sieve.CountPrimes();
-				}
-				{
-					using Sieve sieve = new Sieve();
-					sieve.Initialize(6, 0);
+					using Sieve sieve = Sieve.Create(i, 0);
 					sieve.RunSieve();
 					sieve.CountPrimes();
 				}
@@ -104,8 +86,7 @@ namespace FastSieve
 
 			while (watch.ElapsedTicks < durationLimit)
 			{
-				using var sieve = new Sieve();
-				sieve.Initialize(method, max);
+				using var sieve = Sieve.Create(method, max);
 				sieve.RunSieve();
 				nint value = sieve.CountPrimes();
 				if (value != 78498)
@@ -117,21 +98,14 @@ namespace FastSieve
 			}
 
 			watch.Stop();
-			string methodName;
-			if (method == 1) methodName = "standard";
-			else if (method == 2) methodName = "optimized";
-			else if (method == 3) methodName = "optimized,vectorized";
-			else if (method == 4 || method == 6) methodName = "optimized,unmanaged";
-			else /*if (method == 5 || method == 7)*/ methodName = "optimized,unmanaged,vectorized";
-			Console.WriteLine($"hamarb123_{ methodName }{ (method < 6 ? "" : ",onebit") };{ passes };{ watch.Elapsed.TotalSeconds };1;algorithm=base,faithful={ (methodName.Contains("vectorized") ? "no" : "yes") },bits={ (method < 6 ? 8 : 1) }");
+			Console.WriteLine(Sieve.Create(method, 0).CreateMessage(passes, watch.Elapsed.TotalSeconds));
 		}
 
 		public static void RunTest(nint max, nint expected)
 		{
 			if (max <= 0X7FFFFFC7L) //array based implementations
 			{
-				var sieve1 = new Sieve();
-				sieve1.Initialize(1, max);
+				var sieve1 = Sieve.Create(1, max);
 				sieve1.RunSieve();
 				nint value1 = sieve1.CountPrimes();
 				if (value1 != expected)
@@ -141,16 +115,14 @@ namespace FastSieve
 			}
 			if (max <= (0X7FFFFFC7L * 2 + 1)) //array based implementations
 			{
-				var sieve2 = new Sieve();
-				sieve2.Initialize(2, max);
+				var sieve2 = Sieve.Create(2, max);
 				sieve2.RunSieve();
 				nint value2 = sieve2.CountPrimes();
 				if (value2 != expected)
 				{
 					Console.WriteLine("Invalid result (method 2) with max " + max.ToString() + ", got " + value2.ToString() + ", expected " + expected.ToString());
 				}
-				var sieve3 = new Sieve();
-				sieve3.Initialize(3, max);
+				var sieve3 = Sieve.Create(3, max);
 				sieve3.RunSieve();
 				nint value3 = sieve3.CountPrimes();
 				if (value3 != expected)
@@ -158,32 +130,28 @@ namespace FastSieve
 					Console.WriteLine("Invalid result (method 3) with max " + max.ToString() + ", got " + value3.ToString() + ", expected " + expected.ToString());
 				}
 			}
-			var sieve4 = new Sieve();
-			sieve4.Initialize(4, max);
+			var sieve4 = Sieve.Create(4, max);
 			sieve4.RunSieve();
 			nint value4 = sieve4.CountPrimes();
 			if (value4 != expected)
 			{
 				Console.WriteLine("Invalid result (method 4) with max " + max.ToString() + ", got " + value4.ToString() + ", expected " + expected.ToString());
 			}
-			var sieve5 = new Sieve();
-			sieve5.Initialize(5, max);
+			var sieve5 = Sieve.Create(5, max);
 			sieve5.RunSieve();
 			nint value5 = sieve5.CountPrimes();
 			if (value5 != expected)
 			{
 				Console.WriteLine("Invalid result (method 5) with max " + max.ToString() + ", got " + value5.ToString() + ", expected " + expected.ToString());
 			}
-			var sieve6 = new Sieve();
-			sieve6.Initialize(6, max);
+			var sieve6 = Sieve.Create(6, max);
 			sieve6.RunSieve();
 			nint value6 = sieve6.CountPrimes();
 			if (value6 != expected)
 			{
 				Console.WriteLine("Invalid result (method 6) with max " + max.ToString() + ", got " + value6.ToString() + ", expected " + expected.ToString());
 			}
-			var sieve7 = new Sieve();
-			sieve7.Initialize(7, max);
+			var sieve7 = Sieve.Create(7, max);
 			sieve7.RunSieve();
 			nint value7 = sieve7.CountPrimes();
 			if (value7 != expected)
